Organized Object View to render under Theme
Object views render in the ThemeView render function. I had objects rendering in the routers file. However, the ThemeView was not in the DOM at the time the ObjectView renders.

We may not need to nest Objects within Themes. And we may remove this later.

I also altered the CSS to not make the room background 100% height. I did this so the objects were placed correctly relative to the room. In the future, we'll need to calculate teh offset relative to the browser size.